More work on teh five birds

+++ src/feedback/feedback.controller.ts
@@ -138,7 +138,7 @@ export class FeedbackController {
       user: from,
       targetUser: to,
       key: 'likeability',
-      type: 'double',
+      type: 'int',
       isRating: true,
       value
     } as CreateFlagDTO;
